left_sidebar: Fix stream names visible between PM and stream header.
When scrolling at 110% zoom, it is possible see streams names between
PM and stream header. This is hard to reproduce.

I was not able to reproduce it locally. I was able to reproduce on
chat.zulip.org, so this fix is based on it. We basically add bottom
padding to PM header. This doesn't seem like should fix the bug,
but it surprisingly does. The other solution would be to position
the stream header 1px higher but I didn't do it to avoid a bad
UX resulting from it.